refactor(claw-server-rust): remove services module
