+++ packages/browseros-agent/apps/claw-server-rust/src/capture/audit.rs
@@ -1,4 +1,5 @@
 use crate::{
+    clock::now_epoch_ms,
     db::audit::{
         AuditDb,
         entities::{
@@ -9,7 +10,6 @@ use crate::{
     },
     error::AppResult,
     ids::DispatchId,
-    services::now_epoch_ms,
 };
 use sea_orm::{
     ActiveValue::{NotSet, Set},
