+++ packages/browseros-agent/apps/claw-server-rust/src/clock.rs
@@ -1,7 +1,3 @@
-pub mod agents;
-pub mod harness;
-pub mod telemetry;
-
 pub(crate) fn now_epoch_ms() -> i64 {
     match std::time::SystemTime::now().duration_since(std::time::UNIX_EPOCH) {
         Ok(duration) => i64::try_from(duration.as_millis()).unwrap_or(i64::MAX),
