+++ packages/browseros-agent/apps/claw-server-rust/src/harness.rs
